+++ jest.config.js
@@ -8,4 +8,5 @@ module.exports = {
   transform: {
     ...tsJestTransformCfg,
   },
+  setupFiles: ["./tests/setupEnv.ts"],
 };